tests/wpt/tests/css/css-overflow/resizer-transform.tentative.html

<!doctype html>
<meta charset=utf-8>
<title>Resizer should account for transforms to decide resize direction</title>
<link rel=author href="mailto:emilio@crisal.io" title="Emilio Cobos Ãlvarez">
<link rel=author href="https://mozilla.org" title="Mozilla">
<link rel=help href="https://bugzilla.mozilla.org/show_bug.cgi?id=1775797">
<style>
#resizeme {
  position: absolute;
  top: 200px;
  left: 200px;
  width: 100px;
  height: 100px;
  overflow: hidden;
  resize: both;
  background-color: green;
  transform-origin: 0 0;
  transform: rotate(90deg);
}
</style>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<div id="resizeme"></div>
<script>
promise_test(async function test() {
  let element = document.getElementById("resizeme");
  let rect = element.getBoundingClientRect();

  // Due to the rotation, the resizer should be at the bottom left.
  await new test_driver.Actions()
    .pointerMove(rect.left + 1, rect.bottom - 1)
    .pointerDown()
    .pointerMove(rect.left + 1, rect.bottom + 50)
    .pointerUp()
    .send();

  // We should've made the element wider due to the rotation.
  assert_greater_than(parseInt(getComputedStyle(element).width, 10), 100, "Element should be wider");
  assert_equals(parseInt(getComputedStyle(element).height, 10), 100, "Element should have the same height");
});
</script>
